+++ CHANGELOG.md
@@ -20,6 +20,8 @@ The types of changes are:
 ### Changed
 
 * Changed behavior of `load_default_taxonomy` to append instead of upsert [#1040](https://github.com/ethyca/fides/pull/1040)
+* Deleting a taxonomy field with children will now cascade delete all of its children as well. [#1042](https://github.com/ethyca/fides/pull/1042)
+
 ### Fixed
 
 * Fixed navigating directly to frontend routes loading index page instead of the correct static page for the route.